// apps/stroltm/ui/src/api/generated/api.ts
export interface ApiInfo {
  /**
   *
   * @type {Array<ManagerInfoInstance>}
   * @memberof ApiInfo
   */
  instances?: Array<ManagerInfoInstance>;
  /**
   *
   * @type {string}
   * @memberof ApiInfo
   */
  latestVersion?: string;
  /**
   *
   * @type {string}
   * @memberof ApiInfo
   */
  updatedAt?: string;
  /**
   *
   * @type {string}
   * @memberof ApiInfo
   */
  version?: string;
}
// ...
export interface ManagerInfoInstance {
  /**
   *
   * @type {ManagerInfoInstanceConfig}
   * @memberof ManagerInfoInstance
   */
  config?: ManagerInfoInstanceConfig;
  /**
   *
   * @type {boolean}
   * @memberof ManagerInfoInstance
   */
  isOnline?: boolean;
  /**
   *
   * @type {string}
   * @memberof ManagerInfoInstance
   */
  lastestOnlineAt?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerInfoInstance
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerInfoInstance
   */
  proxyName?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerInfoInstance
   */
  startedAt?: string;
  /**
   *
   * @type {ManagerInfoInstanceTaskStatus}
   * @memberof ManagerInfoInstance
   */
  taskStatus?: ManagerInfoInstanceTaskStatus;
  /**
   *
   * @type {string}
   * @memberof ManagerInfoInstance
   */
  version?: string;
}
/**
 *
 * @export
 * @interface ManagerInfoInstanceConfig
 */
export interface ManagerInfoInstanceConfig {
  /**
   *
   * @type {boolean}
   * @memberof ManagerInfoInstanceConfig
   */
  isInitialized?: boolean;
  /**
   *
   * @type {string}
   * @memberof ManagerInfoInstanceConfig
   */
  updatedAt?: string;
}
/**
 *
 * @export
 * @interface ManagerInfoInstanceTaskStatus
 */
export interface ManagerInfoInstanceTaskStatus {
  /**
   *
   * @type {boolean}
   * @memberof ManagerInfoInstanceTaskStatus
   */
  isInitialized?: boolean;
  /**
   *
   * @type {string}
   * @memberof ManagerInfoInstanceTaskStatus
   */
  updateRequestedAt?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerInfoInstanceTaskStatus
   */
  updatedAt?: string;
}